requirements.txt
sentence-transformers==2.6.1
langchain-huggingface>=0.0.10
streamlit
langchain-core
langchain-community
psycopg2-binary
python-dotenv
pandas
openai
typing-extensions
langgraph
pytest
ruff
isort
black
numpy==1.26.4
pytest-cov
